[BREAKING CHANGE] Save proof steps as Tikz rather than JSON.
To convert proof files (*.zxp) from JSON to Tikz format, cherry-pick this commit (but NOT the subsequent commit). Run ZXLive, open the proof, and save it again (to be on the safe side, as another file). Note that the saved proof will not be loadable in ZXLive until the subsequent commit is also cherry-picked.